pyproject.toml
[project]
name = "style-transfer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow>=12.0.0",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "wandb>=0.22.2",
]
